pyproject.toml
[project]
name = "raspi"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "cgsensor>=1.2",
    "mh-z19>=3.1.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
